refactor: moved the files to appropriate root
