+++ .github/workflows/crawl.yaml
@@ -24,11 +24,11 @@ jobs:
     - name: Test crawler and indexer
       run: |
         source .env/bin/activate
-        python3 -m phantom.phantom --num_threads 2 --urls "https://github.com/AnsahMohammad" "https://github.com/AnsahMohammad" --sleep 6
+        python3 -m phantom.main --num_threads 2 --urls "https://github.com/AnsahMohammad" "https://github.com/AnsahMohammad" --sleep 6
 
         python3 -m nltk.downloader stopwords
         python3 -m nltk.downloader punkt
 
-        python3 -m phantom.phantom_indexing
+        python3 -m phantom.core.indexer
 
         echo "Crawling completed"
