+++ local_search.sh
@@ -5,7 +5,9 @@ source .env/bin/activate
 pip install -r requirements.txt
 clear
 echo "Installation done"
-python3 -m phantom.phantom_indexing
+sleep 2
+python3 -m phantom.core.indexer
 echo "Phantom Indexing done"
+sleep 2
 clear
-python3 -m phantom.query_engine
+python3 -m phantom.core.query_engine
