+++ phantom/core/query_engine.py
@@ -1,6 +1,6 @@
 import json
 from collections import Counter
-from .utils.logger import Logger
+from ..utils.logger import Logger
 import os
 from supabase import create_client, Client
