+++ phantom/requirements.txt
@@ -1,3 +1,4 @@
 bs4==0.0.2
 nltk==3.8.1
 requests==2.31.0
+supabase==2.4.5
